move report_to_cxi to lib
